WIP - unittests
Updated expected values in unittests to correct ones. This resulted in
passing test_smime_sign. Repaired the case in which we need to sign with
passphrase protected cert, which was not loaded correctly due to
explicitly setting the password=None in load function. test_smime_key_cert_together_passphrase still fails, but not not because of the error, rather because the expected value. That could be potentionally right, but I was not able to assure that is the case.