WIP -Encryption
Library has build-in method for encryption, this has been added to the
if encrypt case. This howerer does not satisfy the tests in previous
form, current standard headers don't use `application/x-pkcs7...`, the
"x-" part is deprecated. Other test values were updated to check for
different lines, but all changes were tested manualy if can be decrypted
and/or verified. Which they did.