Allow Certificate Chains in S/MIME Signatures
For some S/MIME signatures, it is not sufficient to include just a single certificate; instead, a chain of certificates must be added to the signature. This is the case whenever intermediate certificates exist between the signing certificate and the root certificate.

With this patch, it is no longer just the first certificate in the chain that is included in the signature, but the entire certificate chain.